+++ tests/accuracy/operators/test_gemm.py
@@ -188,7 +188,7 @@ def generate_random_list(length, total_sum):
 
     diff = total_sum - sum(lst)
     lst[-1] += diff
-    return torch.Tensor(lst).to(torch.int64)
+    return torch.Tensor(lst).to(torch.int32)
 
 
 def generate_quant_group_linear_data(
@@ -243,7 +243,7 @@ def generate_quant_group_linear_data(
         pytest.param(
             torch.randn(size=(256, 128), dtype=dtype),
             torch.randn(size=(1, 128, 64), dtype=dtype),
-            torch.tensor([256], dtype=torch.int64),
+            torch.tensor([256], dtype=torch.int32),
             False,
             id=f"single_group_fp={'bf16' if dtype is torch.bfloat16 else 'fp16'}",
         )
@@ -263,7 +263,7 @@ def generate_quant_group_linear_data(
         pytest.param(
             torch.randn(size=(256, 128), dtype=dtype),
             torch.randn(size=(4, 96, 128), dtype=dtype),
-            torch.tensor([48, 80, 64, 64], dtype=torch.int64),
+            torch.tensor([48, 80, 64, 64], dtype=torch.int32),
             True,
             id=f"trans_weight_uneven_fp={'bf16' if dtype is torch.bfloat16 else 'fp16'}",
         )
@@ -417,7 +417,7 @@ def test_group_linear_two_groups_single_call(dtype, trans_weight):
         weight = torch.stack([w0, w1], dim=0)
         ref = torch.cat([x0 @ w0, x1 @ w1], dim=0)
 
-    group_list = torch.tensor([m0, m1], device=device, dtype=torch.int64)
+    group_list = torch.tensor([m0, m1], device=device, dtype=torch.int32)
 
     op = MojoGroupLinear(weight=weight, trans_weight=trans_weight)
     out = op(x, group_list)
